+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## v4.4
 - Replaced Cookie login with Email:Password
--
+- Fixed 4 bugs
 
 ## [v4.3](https://github.com/techtanic/Discounted-Udemy-Course-Enroller/releases/tag/v4.3) - 2021-05-20